alp and helins code merge off app.js
