+++ client/package.json
@@ -14,7 +14,7 @@
     "framer-motion": "^10.12.7",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.11.0",
+    "react-router-dom": "^6.11.1",
     "react-scripts": "^5.0.1",
     "sass": "^1.62.1",
     "web-vitals": "^2.1.4"
