--- client/src/app_frontend.js
@@ -1,29 +1,0 @@
-import { useState } from "react";
-import Header from "./components/Layout/Header";
-import Products from "./components/Products/Products";
-import Cart from "./components/Cart/Cart";
-import CartProvider from "./store/cartProvider";
-//import LoginSignup from "./components/Login/Login";
-function App() {
-  const [cartIsShown, setCartIsShown] = useState(false);
-
-  const showCartHandler = () => {
-    setCartIsShown(true);
-  };
-
-  const hideCartHandler = () => {
-    setCartIsShown(false);
-  };
-
-  return (
-    <CartProvider>
-      {cartIsShown && <Cart onClose={hideCartHandler} />}
-      <Header onShowCart={showCartHandler} />
-      <main>
-        <Products />
-      </main>
-    </CartProvider>
-  );
-}
-
-export default App;
